requirements.txt
fastapi==0.75.2
python-dotenv==0.20.0
uvicorn[standard]==0.17.6
PyJWT==2.3.0
python-multipart==0.0.5
APScheduler===3.9.1
bleach==5.0.0
SQLAlchemy==2.0.3
PyMySQL==1.0.2
mysqlclient==2.1.1

